Revert "Merge branch 'Request' of https://github.com/Skiddygoat15/Neighbourhood-Walk into Request"
This reverts commit 80b8e381c53cc11974bdd1fa21ef125555be5db7, reversing
changes made to 7e5140ddc4f91b80df53e0449133210d15b50130.

+++ src/main/java/com/comp5703/Neighbourhood/Walk/Controller/RequestController.java
@@ -76,7 +76,7 @@ public class RequestController {
      * @return
      */
     @PostMapping("/{requestId}/accept")
-    public ResponseEntity<?> acceptRequest(@PathVariable int requestId, @RequestParam int walkerId) {
+    public ResponseEntity<?> acceptRequest(@PathVariable int requestId, @RequestParam long walkerId) {
         if (requestService.acceptWalkerRequest(requestId, walkerId) == null){
             return new ResponseEntity<>("The request has been accepted by some walker.", HttpStatus.BAD_REQUEST);
         }
@@ -94,7 +94,7 @@ public class RequestController {
      * @return
      */
     @PostMapping("/{requestId}/reject")
-    public ResponseEntity<?> rejectRequest(@PathVariable int requestId, @RequestParam int walkerId) {
+    public ResponseEntity<?> rejectRequest(@PathVariable int requestId, @RequestParam long walkerId) {
         if (requestService.rejectWalkerRequest(requestId, walkerId) == null){
             return new ResponseEntity<>("The request has been rejected by this walker", HttpStatus.BAD_REQUEST);
         }
